requirements.txt
streamlit==1.47.0
chromadb==1.0.15
sentence-transformers==5.0.0
openai==1.97.0
tavily-python==0.7.10
python-dotenv==1.1.1
pydantic==2.11.7
requests==2.32.4
tqdm==4.67.1
pandas==2.3.1
scikit-learn==1.7.1
numpy==2.3.1
